+++ libft/srcs/ft_pow.c
@@ -6,13 +6,13 @@
 /*   By: mykman <mykman@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/28 02:02:24 by mykman            #+#    #+#             */
-/*   Updated: 2021/05/28 17:05:52 by mykman           ###   ########.fr       */
+/*   Updated: 2021/06/15 14:59:04 by mykman           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-unsigned long	ft_pow(unsigned long n, int pow)
+unsigned long long	ft_pow(unsigned long long n, int pow)
 {
 	if (pow <= 0)
 		return (1);
